refactor(invites): give candidates fan-out its own server fn, revert leaderboard bypass flag
Architectural call from review: invite candidacy and the public
leaderboard have different definitions of "eligible". The previous fix
(`bypassHeatBasedDivisionFilter` flag on `getCompetitionLeaderboard`)
cleared the symptom but coupled them through a growing list of
conditional flags.

This change:
- Adds `getCandidatesForSourceComp({ competitionId, divisionId })` in
  src/server/competition-invites/candidates.ts. Owns its own SQL against
  competition_registrations joined to users (and scaling_levels for the
  label), filtered by eventId, divisionId, and status != "removed". No
  track-workout, heat, event-status, or publication gating.
- `getChampionshipRoster` now fans out to `getCandidatesForSourceComp`
  instead of `getCompetitionLeaderboard`. The previous email-hydration
  second pass is gone — the candidates fn returns email inline.
- Reverts the `bypassHeatBasedDivisionFilter` flag from
  `getCompetitionLeaderboard`.
- Replaces `roster-heat-filter-bypass.test.ts` with
  `roster-candidates-wiring.test.ts`, which forbids the leaderboard
  call and pins the new contract.
- Adds `candidates.test.ts` covering the bug scenario directly.

TDD red→green: candidates.test.ts was authored against the missing
file and ran red before the implementation landed.

Tradeoff: the candidates path no longer benefits from leaderboard score
ranking; rows are stable-ordered by registeredAt. Score-based ranking
inside the candidate set is a follow-up — when it lands it'll live in
candidates.ts, not as another flag on the leaderboard.

Tracking: broadcast-4m8
