+++ lat.md/competition-invites.md
@@ -58,7 +58,7 @@ Loader gates on `MANAGE_COMPETITIONS` via [[apps/wodsmith-start/src/server-fns/c
 
 `getChampionshipRoster({ championshipId, divisionId })` in [[apps/wodsmith-start/src/server/competition-invites/roster.ts]] returns an ordered `RosterRow[]` for a championship + division. Invite-state columns are `null` in Phase 1.
 
-The function loads all sources, delegates to [[apps/wodsmith-start/src/server/competition-leaderboard.ts#getCompetitionLeaderboard]] for single-comp sources and [[apps/wodsmith-start/src/server/series-leaderboard.ts#getSeriesLeaderboard]] for series sources, resolves each source's division mapping, and flags rows below each source's spot allocation as `belowCutoff`. Pure helpers (`parseDivisionMappings`, `resolveSourceDivisionId`, `resolveSpotsForDivision`, `aggregateQualifyingRows`) are extracted and unit-tested directly so the cutoff + skip-already-qualified logic can be exercised without mocking the leaderboard stack. The server fn `getChampionshipRosterFn` gates this with `MANAGE_COMPETITIONS` on the championship.
+The function loads all sources, delegates to [[apps/wodsmith-start/src/server/competition-invites/candidates.ts#getCandidatesForSourceComp]] for single-comp sources and [[apps/wodsmith-start/src/server/series-leaderboard.ts#getSeriesLeaderboard]] for series sources, resolves each source's division mapping, and flags rows below each source's spot allocation as `belowCutoff`. Pure helpers (`parseDivisionMappings`, `resolveSourceDivisionId`, `resolveSpotsForDivision`, `aggregateQualifyingRows`) are extracted and unit-tested directly so the cutoff + skip-already-qualified logic can be exercised without mocking the candidates stack. The server fn `getChampionshipRosterFn` gates this with `MANAGE_COMPETITIONS` on the championship.
 
 Series sources have two qualifier tiers handled by `loadSeriesDirectRows` and `loadSeriesGlobalRows`. Direct qualifiers come from each comp's own leaderboard — the source's `directSpotsPerComp` × number of comps in the series. Globals come from the series-aggregate leaderboard with direct qualifiers filtered out. The per-comp division for each comp is resolved via [[apps/wodsmith-start/src/db/schemas/series.ts#seriesDivisionMappingsTable]]; the resulting source cutoff expands to `directCount + globalSpots` so both tiers count as qualified for the `belowCutoff` waterline.
 
@@ -66,7 +66,13 @@ Series sources have two qualifier tiers handled by `loadSeriesDirectRows` and `l
 
 ADR-0012 Phase 6: per-leaderboard cutoffs in `getChampionshipRoster` no longer read source-level scalars directly. Instead the function loads the championship's allocation rows once via [[apps/wodsmith-start/src/server/competition-invites/allocations.ts#listAllocationsForChampionship]], computes a per-source resolved map via [[apps/wodsmith-start/src/server/competition-invites/allocations.ts#resolveSourceAllocations]] (cached per source), and uses the resolved value for `(sourceId, championshipDivisionId)` to truncate each `(sourceComp × sourceDivision)` leaderboard. The source's `globalSpots` / `directSpotsPerComp` act as defaults when no allocation row exists. Source-side divisions are mapped to championship divisions via the source's `divisionMappings` JSON, falling back to case-insensitive label match; an unmappable division returns `null` (no truncation) rather than risk dropping qualifiers. Series sources keep the full leaderboard for now — splitting per-comp direct vs. global tiers under per-division overrides is a follow-up TODO documented inline.
 
-The fan-out passes `bypassHeatBasedDivisionFilter: true` to [[apps/wodsmith-start/src/server/competition-leaderboard.ts#getCompetitionLeaderboard]] for every `(sourceComp × sourceDivision)` read. Heat assignments are an in-person scoring detail (which lane/time block an athlete runs in), not a qualifier-eligibility signal — a source-comp division with registered athletes but no heat rows referencing it must still surface its athletes as candidates. Without the bypass, the leaderboard's heat-based fallback empties the events list when no heat references the requested division, returning an empty leaderboard and silently dropping that division from the championship roster. Event-division mappings and registration-level division filters still apply. See `docs/bugs/0001-invite-candidates-missing-divisions.md`.
+The fan-out calls [[apps/wodsmith-start/src/server/competition-invites/candidates.ts#getCandidatesForSourceComp]] for every `(sourceComp × sourceDivision)` read instead of going through the public leaderboard. Earlier the roster shared `getCompetitionLeaderboard`, which inherits the leaderboard's heat-based event filter and division-publication gates — a source-comp division with registered athletes but no heat rows referencing it would silently disappear from the championship roster. The dedicated query owns its own SQL against `competitionRegistrationsTable` so the two paths can evolve independently. See `docs/bugs/0001-invite-candidates-missing-divisions.md`.
+
+## Candidates query
+
+[[apps/wodsmith-start/src/server/competition-invites/candidates.ts#getCandidatesForSourceComp]] is the source-of-truth read for the championship roster's per-`(sourceComp, sourceDivision)` fan-out. It returns active registrations as `CandidateEntry[]` with athlete name + email already hydrated, ordered by `registeredAt`.
+
+The query joins `competition_registrations` → `users` (inner) → `scaling_levels` (left), filtered by `eventId`, `divisionId`, and `status ≠ "removed"`. There is no track-workout, heat, event-status, or publication gating — every active registration in the division surfaces. This is the architectural fix for the missing-divisions bug: invite candidacy is "is this athlete registered in the division" not "did this athlete score on a published event in a heat." Score-based ranking inside the candidate set is a follow-up; today rows are stable-ordered by `registeredAt`.
 
 ## Organizer route shell
